Dialog refactor
