+++ data/text/dialog_lookup_table.py
@@ -1,12 +1,11 @@
 from functools import partial
 
-from pygame import display, time, mixer, KEYDOWN, K_DOWN, K_UP, K_w, K_s, K_k, event, K_RETURN, K_j
+from pygame import display, time, mixer, KEYDOWN, K_DOWN, K_UP, K_w, K_s, K_k, K_RETURN, K_j
 from pygame.event import get, pump
 from pygame.time import get_ticks
 
-from data.text.dialog import blink_yes_confirmation, blink_no_confirmation
-from src.common import play_sound, confirmation_sfx, special_item_sfx, CONFIRMATION_BACKGROUND_PATH, CONFIRMATION_YES_BACKGROUND_PATH, \
-    menu_button_sfx, CONFIRMATION_NO_BACKGROUND_PATH, BRECCONARY_WEAPONS_SHOP_PATH, convert_to_frames_since_start_time
+from data.text.dialog import confirmation_prompt, get_inn_intro
+from src.common import play_sound, special_item_sfx, BRECCONARY_WEAPONS_SHOP_PATH, convert_to_frames_since_start_time
 from src.config import MUSIC_ENABLED, TILE_SIZE
 from src.game_functions import draw_all_tiles_in_current_map
 from src.items import weapons, armor, shields
@@ -117,69 +116,19 @@ class DialogLookup:
 
     def tantegel_throne_room_roaming_guard(self):
         player_please_save_the_princess = f"{self.player.name}, please save the Princess."
-        self.confirmation_prompt("Dost thou know about Princess Gwaelin?",
-                                 yes_path_function=partial(self.command_menu.show_line_in_dialog_box, player_please_save_the_princess),
-                                 no_path_function=partial(self.command_menu.show_text_in_dialog_box,
-                                                          ("Half a year now hath passed since the Princess was kidnapped by the enemy.",
-                                                           "Never does the King speak of it, but he must be suffering much.",
-                                                           player_please_save_the_princess),
-                                                          drop_down=False, drop_up=False,
-                                                          skip_text=self.command_menu.skip_text))
-
-    def confirmation_prompt(self, prompt_line, yes_path_function, no_path_function, finally_function=None, skip_text=False):
-        self.command_menu.show_line_in_dialog_box(prompt_line, skip_text=True)
-        self.command_menu.window_drop_down_effect(4, 3, 5, 2)
-        self.command_menu.create_window(5, 2, 4, 3, CONFIRMATION_BACKGROUND_PATH)
-        display.flip()
-        play_sound(confirmation_sfx)
-        blinking = True
-        blinking_yes = True
-        while blinking:
-            if blinking_yes:
-                blink_yes_confirmation(self.command_menu)
-            else:
-                blink_no_confirmation(self.command_menu)
-            if skip_text:
-                play_sound(menu_button_sfx)
-                yes_path_function()
-                blinking = False
-            for current_event in get():
-                if current_event.type == KEYDOWN:
-                    if current_event.key in (K_DOWN, K_UP, K_w, K_s):
-                        if blinking_yes:
-                            self.command_menu.create_window(5, 2, 4, 3, CONFIRMATION_NO_BACKGROUND_PATH)
-                            blinking_yes = False
-                        else:
-                            self.command_menu.create_window(5, 2, 4, 3, CONFIRMATION_YES_BACKGROUND_PATH)
-                            blinking_yes = True
-                    elif (blinking_yes and current_event.unicode in ('\r', 'k')) or current_event.unicode == 'y':
-                        self.command_menu.create_window(5, 2, 4, 3, CONFIRMATION_YES_BACKGROUND_PATH)
-                        play_sound(menu_button_sfx)
-                        event.pump()
-                        yes_path_function()
-                        blinking = False
-                    elif (not blinking_yes and current_event.unicode in ('\r', 'k')) or current_event.unicode in ('n', 'j'):
-                        self.command_menu.create_window(5, 2, 4, 3, CONFIRMATION_NO_BACKGROUND_PATH)
-                        play_sound(menu_button_sfx)
-                        event.pump()
-                        no_path_function()
-                        blinking = False
-            event.pump()
-
-        if finally_function is not None:
-            finally_function()
-
-    @staticmethod
-    def get_inn_intro(inn_cost):
-        return "Welcome to the traveler's Inn.\n" \
-               f"Room and board is {inn_cost} GOLD per night.\n" \
-               "Dost thou want a room?"
+        confirmation_prompt(self.command_menu, "Dost thou know about Princess Gwaelin?",
+                            yes_path_function=partial(self.command_menu.show_line_in_dialog_box, player_please_save_the_princess),
+                            no_path_function=partial(self.command_menu.show_text_in_dialog_box,
+                                                     ("Half a year now hath passed since the Princess was kidnapped by the enemy.",
+                                                      "Never does the King speak of it, but he must be suffering much.",
+                                                      player_please_save_the_princess),
+                                                     drop_down=False, drop_up=False,
+                                                     skip_text=self.command_menu.skip_text))
 
     def check_buy_weapons_armor(self, current_store_inventory, static_store_image):
-        self.confirmation_prompt(weapons_and_armor_intro,
-                                 # TODO(ELF): Add store inventory to yes_path_function.
-                                 yes_path_function=partial(self.open_store_inventory, current_store_inventory, static_store_image),
-                                 no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Please, come again."))
+        confirmation_prompt(self.command_menu, weapons_and_armor_intro,
+                            yes_path_function=partial(self.open_store_inventory, current_store_inventory, static_store_image),
+                            no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Please, come again."))
 
     def open_store_inventory(self, current_store_inventory, static_store_image):
         self.command_menu.show_line_in_dialog_box("What dost thou wish to buy?", skip_text=True)
@@ -238,15 +187,15 @@ class DialogLookup:
                 old_item_cost = self.shopkeeper_buy_old_item(old_item_cost, self.player.armor, armor)
             elif selected_item_type == 'shield':
                 old_item_cost = self.shopkeeper_buy_old_item(old_item_cost, self.player.shield, shields)
-            self.confirmation_prompt("Is that Okay.?",
-                                     yes_path_function=partial(self.complete_transaction, selected_item, current_store_inventory, old_item_cost),
-                                     no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Oh, yes? That's too bad."))
+            confirmation_prompt(self.command_menu, "Is that Okay.?",
+                                yes_path_function=partial(self.complete_transaction, selected_item, current_store_inventory, old_item_cost),
+                                no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Oh, yes? That's too bad."))
         else:
             self.command_menu.show_line_in_dialog_box("Sorry.\n"
                                                       "Thou hast not enough money.")
-        self.confirmation_prompt("Dost thou wish to buy anything more?",
-                                 yes_path_function=partial(self.open_store_inventory, current_store_inventory, static_store_image),
-                                 no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Please, come again."))
+        confirmation_prompt(self.command_menu, "Dost thou wish to buy anything more?",
+                            yes_path_function=partial(self.open_store_inventory, current_store_inventory, static_store_image),
+                            no_path_function=partial(self.command_menu.show_line_in_dialog_box, "Please, come again."))
 
     def shopkeeper_buy_old_item(self, old_item_cost, old_item, old_item_lookup_table):
         if old_item:
@@ -271,13 +220,12 @@ class DialogLookup:
         self.command_menu.show_line_in_dialog_box("I thank thee.")
 
     def check_stay_at_inn(self, inn_cost):
-        self.confirmation_prompt(self.get_inn_intro(inn_cost),
-                                 yes_path_function=partial(self.check_money, inn_cost),
-                                 no_path_function=partial(self.command_menu.show_line_in_dialog_box,
-                                                          "Okay.\n"
-                                                          "Good-bye, traveler.",
-                                                          skip_text=self.command_menu.skip_text),
-                                 skip_text=self.command_menu.skip_text)
+        confirmation_prompt(self.command_menu, get_inn_intro(inn_cost),
+                            yes_path_function=partial(self.check_money, inn_cost),
+                            no_path_function=partial(self.command_menu.show_line_in_dialog_box,
+                                                     "Okay.\n"
+                                                     "Good-bye, traveler.",
+                                                     skip_text=self.command_menu.skip_text), skip_text=self.command_menu.skip_text)
 
     def check_money(self, inn_cost):
         if self.player.gold >= inn_cost:
